Add more descriptive var name and comment
So we've been searching for why the new animation is not working.
This doesn't solve it, but I think that I've run into the reason

Our original animation had a parent -> child -> child -> child
joint relationship and worked fine, however the great grand child bone
was NOT being animated

In or second animation it IS being animated, and it isn't working.
So it looks like great grand parent relationships are broken.

If we're correct, this means that our recursion isn't actually working. It's only working
out for the first three bones.

So the next action to take is two create a simple test model with
a parent->child->child->child relationship, animate the bones, and get
the tests to pass